requirements.txt
kivy>=2.1.0
kivymd>=1.1.1
buildozer>=1.4.0
cython>=0.29.33
python-for-android>=2023.6.11
pyjnius>=1.4.2